esta vivooooo

+++ trunk/src/CajaVelocidades.java
@@ -13,7 +13,7 @@ public class CajaVelocidades extends Autoparte{
 		
 		relaciones[0] = 0;
 		for(int i = 1; i < (cantidadCambios + 1); i++) 
-			relaciones[i] = ((cantidadCambios+1-i)/cantidadCambios ); 
+			relaciones[i] = ((double)(cantidadCambios+1-i)/cantidadCambios ); 
 	}
